Smash bros camera with linear interpolation
Implemented a camera which averages all the players and balls in the scene. Camera was too jerky, so implemented linear interpolation. Works great, except character looks a bit jerky if they're at the edge of the screen which needs fixing. Also breaks if a ball is deleted (which should be an easy fix). PEACE!