Attack orb raycasting MOFO!!!
Fixed the annoying effing attack orb clipping through targets!!! It checks its position against its previous position and casts rays backwards. Everything in the ray is then checked in order. The first thing hits react. TODO: Add a second layer of checking for spherical collision, as the ray is a thin line and does not account for the diameter of the attack orb.